refactor(branded): split brand-icon registry from component for fast refresh
